feat(pegout): add initial cache impl
